Rename cacheMiddleware to listenerMiddleware
